persona module added

+++ src/main/java/com/modular/modules/Persona/Service/PersonaService.java
@@ -1,36 +1,32 @@
 package com.modular.modules.Persona.Service;
-import com.modular.modules.Persona.Dto.CreatePersonaDto;
 
 import java.util.List;
 import java.util.Optional;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import com.modular.modules.Persona.Dto.CreatePersonaDto;
 import com.modular.modules.Persona.Entity.Persona;
 import com.modular.modules.Persona.PersonaRepository;
+import org.springframework.http.HttpStatus;
 
 @Service
 public class PersonaService {
 
-    @Autowired
-    private PersonaRepository personaRepository;
-    private static final Logger logger = LoggerFactory.getLogger(PersonaService.class);
+    private final PersonaRepository personaRepository;
 
-    public List<Persona> getAllPersonas() {
-        List<Persona> personas = personaRepository.findAll();
-        if (personas.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontraron personas.");
-        }
-        return personas;
+    public PersonaService(final PersonaRepository personaRepository) {
+        this.personaRepository = personaRepository;
     }
 
-    public Optional<Persona> getPersonaById(Long id) {
-        return personaRepository.findById(id);
+    public List<Persona> getAllPersonas() {
+        return personaRepository.findAll();
+    }
+
+    public Persona getPersonaById(Long id) {
+        return personaRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Persona no encontrada"));
     }
 
     public Persona createPersona(CreatePersonaDto dto) {
@@ -41,13 +37,16 @@ public class PersonaService {
     }
 
     public Persona updatePersona(Long id, Persona personaDetails) {
-        Persona persona = personaRepository.findById(id).orElseThrow();
+        Persona persona = this.getPersonaById(id);
         persona.setNombre(personaDetails.getNombre());
         persona.setApellido(personaDetails.getApellido());
         return personaRepository.save(persona);
     }
 
     public void deletePersona(Long id) {
+        if (!personaRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Persona no encontrada");
+        }
         personaRepository.deleteById(id);
     }
 }
